Fix two chain-frg bugs found via live FRG devnet testing
submit_settlement_batch passed the "sett" selector to FrgWalletClient::
call_contract's separate `function` param while also passing the built
calldata as `call_data`. frg-wallet's /contracts/call prioritizes
call_data_hex over function whenever both are set (see cmd/frg-wallet/
main.go's contractCallData), so the selector was silently discarded and
the raw body -- starting with pre_root, not "sett" -- got sent as literal
calldata. FRG read garbage bytes as the function name and the call failed.
Now the selector is prepended into the calldata itself, matching what
FrgWalletClient::call_contract's own docs already said should happen.

get_node_stake (and the FrgClient::validator_bond it wraps) parsed a
validator's bond into a u64. Real FRG bonds are quanta (1 FRG = 10^18
quanta), and the protocol minimum bond alone is 1,000 FRG = 10^21 quanta,
~54x past u64::MAX -- so this failed outright for every real validator.
FrgClient::validator_bond now returns u128 (the wire-correct precision);
only get_node_stake, stuck with u64 by the shared ChainAdapter trait
signature, saturates at the boundary, documented as lossy rather than
silently wrong.

Both confirmed against a live FRG devnet (deployed contracts/frg/
settlement, ran submit_settlement_batch with a real proof end-to-end,
is_trade_settled flipped from false to true).

+++ crates/chain-frg/src/client.rs
@@ -21,12 +21,18 @@ impl FrgClient {
         Ok(Self { inner })
     }
 
-    pub async fn validator_bond(&self, pubkey: OnChainAccount) -> Result<u64, String> {
+    // u128, not u64: FRG bonds are quanta (1 FRG = 10^18 quanta), and the
+    // protocol minimum bond alone is 1,000 FRG = 10^21 quanta -- already
+    // ~54x past u64::MAX (~1.8*10^19, i.e. ~18.4 FRG). u64 can't represent
+    // any real validator's bond. Confirmed live against a real devnet:
+    // parsing this as u64 failed outright for the genesis validator's
+    // bond. u128 covers the full range with room to spare.
+    pub async fn validator_bond(&self, pubkey: OnChainAccount) -> Result<u128, String> {
         let entry = self.find_validator(pubkey).await?;
         match entry {
             Some(v) => v
                 .bond
-                .parse::<u64>()
+                .parse::<u128>()
                 .map_err(|e| format!("chain-frg: unparsable bond {:?}: {e}", v.bond)),
             None => Ok(0),
         }
